parent blocks: link to polyblock head for children
If a parent refers to a child block and that child block is a polyblock
chain, link to the first block in the chain instead of the last one.
Previously we unconditionally updated the HTML ID of a child block
whenever we found a source code block with the child's name; now we only
update the HTML ID if we are doing it for the first time, so that we
always only refer to the first block of a polyblock.

For the record, before 9284e5b (give all source code blocks an HTML ID,
2023-05-26), we used to give multiple source code blocks the same HTML
ID. This was wrong but it "worked" because the web browser would
automatically always jump to the first occurrence of the HTML ID.