css: hide horizontal scrollbar
The horizontal scrollbar can appear if we have text in <pre> tags that
are too long. In this scenario, if the window is resized to be narrow
enough, horizontal scrollbars will appear and we'll be able to scroll
right, which would make the main body text overlap with the table of
contents on the left. This breaks the UI.

The fix is to just disable horizontal scrollbars altogether.