make external link detection more robust
In 6d9bf91 (move developer guide to separate file, 2023-08-09), we
created the developer guide and one of the links to the monoblock
examples became an external link (due to the target section being in
README.org, not in developer-guide.org), when we updated the Org mode
bit from

    [[* Monoblock (evaluation result's value)]]

to

    [[file:README.org::#h-Monoblock--evaluation-result-s-value][Monoblock (evaluation resut's value)]].

The link changed from "#h-Monoblock-..." in the HTML to
"README.html#h-Monoblock-...". This link was not handled properly
because of our custom JS which treated it to be an internal link. This
was because our external link detection was not very good.

Fix this by making the external link detection more robust.