css: use tango theme
We now use an explicit theme and no longer use a big, black-box hack to
make CSS color generation work in batch mode in emacs. The only tradeoff
is that we can no longer run the CSS generation rule in
non-interactive (batch) mode. This is OK because we rarely need to
change the default theme that we use.

Also make the CSS generation no longer be called whenever
developer-guide.org is updated. Instead, CSS generation must be done
manually every time we change the theme.