css: add enough padding on the left
Some citation sytles like apa.csl result in a citation with a negative
"text-indent" property to create the illusion of a handing indent.
However this does not play well when we try to highlight a selection box
around the entire entry, because the negative text-indent will make the
text bleed over to the left.

Make the padding large enough on the left so that we'll still enclose
around the negative text-indent.