css: use the same font for all "pre" elements
Previously, we never set the font for this element so it would default
to "monospace", which could be different from "var(--font-mono)".

This resulted in example blocks being rendered in a different font than
source code blocks, resulting in a different size as well.

Use the same font for all "pre" elements.